AI-Powered Financial Annual Report Intelligence System/Application Bundle/requirements.txt
pymupdf
pydantic
pytesseract
pillow